requirements.txt
ffmpeg~=1.4
PyYAML~=5.4.1
aiohttp~=3.7.4.post0
mutagen~=1.45.1
pytube~=10.6.1
Telethon~=1.21.1
ffmpeg-python~=0.2.0